Refactor handling of separated records in streams
Much of the output desired from `git` comes in the form of records
separated by newlines or by NUL characters.  In either case, it is nice
to process record-by-record while still plugging into the NodeJS stream
architecture.  `SeparatedRecordConsumer` provides that bridge, emitting
a `record` event for each record streamed to it.